Create README.md
This repository is made to facilitate some of the more interesting projects in Mathematics and Physics which I have made throughout my Junior Spring semester of Undergraduate studies, while obtaining a degree in Mathematics and Physics. It contains some of the more challenging graphs and diagrams, interesting proofs and detailed explanations of problems regularly found in undergraduate textbooks.

Remark: "Interference.png" is a graph obtained for an experiment which was done on the Double Slit interference of a red-light laser. The data was obtained from a TeachSpin apparatus, with a modified automatized data acquisition system which I built, and superimposed against a Fraunhofer model considering light of wavelength 675 nm. The fact that the two graphs are concordant shows that the Fraunhofer model holistically represents double-slit interference, since 675 nm is the nominal working wavelength of the laser used in the experiment.